requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
asyncio==4.0.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
environs==14.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
marshmallow==4.0.1
ollama==0.6.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.10
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
ruff==0.14.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
